Token base class, Vault
Move shared functionality to a base class Token.
It's an ERC20 with additional data that holds addresses of every account
that has ever held the token.
This is useful for migration so accounts can be iterated.

The Vault is rudmentary, WIP.
Will accept RewardToken and also emit RewardToken rewards to stakers.
Itself is an ERC20 so that users may trade their staked tokens.