Added more interfaces, deployment is a bitch
Turns out you need to flatten your contracts to have it verified. Makes
sense in hindsight - it's just not what people mention when learning
this stuff.

Commented out voting for now, going to see if this version without
voting can run on testnet and then add it later if there's enough time.

Otherwise I'll add it after the hackathon and before mainnet launch.